add ClassItem and ScheduleData to Types file
